polymarket/requirements.txt
requests
pandas
python-dateutil
openpyxl
